requirements.txt
playwright==1.44.0
gspread==6.1.2
google-api-python-client==2.130.0
google-auth==2.29.0
python-dotenv==1.0.1
rapidfuzz==3.9.3
APScheduler==3.10.4
